feat(desktop): Lexical 编辑器壳层替换 contenteditable 核心
- 新增 ComposerLexicalInput，composer-rich-input 改为薄 re-export
- 编辑器变更经 editorStateToRichSegments 驱动 segments，移除 DOM 双向同步
